FAB-12548 Update node, stable versions in CI
Update Nodejs and STABLE version to latest release-1.3
versions in CI pipeline script. Also, add java chaincode
tests in byfn and eyfn.

+++ scripts/Jenkins_Scripts/CI_Script.sh
@@ -1,17 +1,12 @@
-#!/bin/bash -e
+#!/bin/bash
 #
 # Copyright IBM Corp All Rights Reserved
 #
 # SPDX-License-Identifier: Apache-2.0
 #
-
-# exit on first error
-
 export BASE_FOLDER=$WORKSPACE/gopath/src/github.com/hyperledger
-export PROJECT_VERSION=1.3.0-stable
 export NEXUS_URL=nexus3.hyperledger.org:10001
 export ORG_NAME="hyperledger/fabric"
-export NODE_VER=8.9.4 # Default nodejs version
 
 # Fetch baseimage version
 curl -L https://raw.githubusercontent.com/hyperledger/fabric/master/Makefile > Makefile
@@ -21,12 +16,7 @@ export OS_VER=$(dpkg --print-architecture)
 echo "-----------> OS_VER" $OS_VER
 export BASE_IMAGE_TAG=$OS_VER-$BASE_IMAGE_VER
 
-# Fetch Go Version from fabric ci.properties file
-curl -L https://raw.githubusercontent.com/hyperledger/fabric/master/ci.properties > ci.properties
-export GO_VER=`cat ci.properties | grep GO_VER | cut -d "=" -f 2`
-echo "-----------> GO_VER" $GO_VER
-
-# Published stable version from nexus
+# Stable version from nexus
 export STABLE_TAG=$OS_VER-$PROJECT_VERSION
 echo "-----------> STABLE_TAG" $STABLE_TAG
 
@@ -35,9 +25,6 @@ Parse_Arguments() {
               case $1 in
                       --env_Info)
                             env_Info
-                            ;;
-                      --SetGopath)
-                            setGopath
                             ;;
                       --pull_Fabric_Images)
                             pull_Fabric_Images
@@ -117,25 +104,18 @@ env_Info() {
 	docker info
 	docker-compose version
 	pgrep -a docker
-	docker images
-	docker ps -a
 }
 
-setGopath() {
-        echo "-----------> set GOPATH"
-	echo
-	export GOPATH=$WORKSPACE/gopath
-	export JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64
-        export PATH=$GOROOT/bin:$GOPATH/bin:/usr/local/bin:/usr/bin:/usr/local/sbin:/usr/sbin:~/npm/bin:/home/jenkins/.nvm/versions/node/v6.9.5/bin:/home/jenkins/.nvm/versions/node/v$NODE_VER/bin:$PATH
-        export GOROOT=/opt/go/go$GO_VER.linux.$OS_VER
-	export PATH=$GOROOT/bin:$PATH
-}
 # Pull Thirdparty Docker images (Kafka, couchdb, zookeeper)
 pull_Thirdparty_Images() {
             for IMAGES in kafka couchdb zookeeper; do
-                 echo "-----------> Pull $IMAGE image"
+                 echo "-----------> Pull $IMAGES image"
                  echo
-                 docker pull $ORG_NAME-$IMAGES:$BASE_IMAGE_TAG
+                 docker pull $ORG_NAME-$IMAGES:$BASE_IMAGE_TAG > /dev/null 2>&1
+                 if [ $? -ne 0 ]; then
+                       echo -e "\033[31m FAILED to download docker images" "\033[0m"
+                       exit 1
+                 fi
                  docker tag $ORG_NAME-$IMAGES:$BASE_IMAGE_TAG $ORG_NAME-$IMAGES
             done
                  echo
@@ -143,11 +123,14 @@ pull_Thirdparty_Images() {
 }
 # pull fabric images from nexus
 pull_Fabric_Images() {
-        setGopath fabric # set gopath
             for IMAGES in peer orderer tools ccenv; do
                  echo "-----------> pull $IMAGES image"
                  echo
-                 docker pull $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG
+                 docker pull $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG > /dev/null 2>&1
+                 if [ $? -ne 0 ]; then
+                       echo -e "\033[31m FAILED to download docker images" "\033[0m"
+                       exit 1
+                 fi
                  docker tag $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG $ORG_NAME-$IMAGES
                  docker tag $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG $ORG_NAME-$IMAGES:$STABLE_TAG
                  docker rmi -f $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG
@@ -158,11 +141,13 @@ pull_Fabric_Images() {
 # pull fabric-ca images from nexus
 pull_Fabric_CA_Image() {
 	echo
-        setGopath fabric-ca
             for IMAGES in ca ca-peer ca-orderer ca-tools; do
                  echo "-----------> pull $IMAGES image"
                  echo
-                 docker pull $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG
+                 docker pull $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG > /dev/null 2>&1
+                 if [ $? -ne 0 ]; then
+                       echo -e "\033[31m FAILED to download docker images" "\033[0m"
+                 fi
                  docker tag $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG $ORG_NAME-$IMAGES
 	         docker tag $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG $ORG_NAME-$IMAGES:$STABLE_TAG
                  docker rmi -f $NEXUS_URL/$ORG_NAME-$IMAGES:$STABLE_TAG
@@ -175,5 +160,8 @@ byfn_eyfn_Tests() {
 	echo
 	echo "-----------> Execute Byfn and Eyfn Tests"
 	./byfn_eyfn.sh
+        if [ $? -ne 0 ]; then
+              echo -e "\033[31m FAILED BYFN Tests" "\033[0m"
+        fi
 }
 Parse_Arguments $@
